reorg and handle NumPy
